refactor: Create `ChatUI` directory, separate components to their own files
- Separate `ModelSelector`, `ChatInput`, and `SettingsPanel` into their own files.
- Rename `ChatUI.tsx` to `Chat.tsx`
- Use `ChatUI.tsx` in `index.tsx`, imports all components and exports the final ui as `ChatUI`
- Update import in `gptchat.astro`
