+++ src/components/ChatUI/index.tsx
@@ -1,0 +1,15 @@
+import { SettingsPanel } from './SettingsPanel';
+import { Chat } from './Chat';
+
+const ChatUI: React.FC = () => {
+  return (
+    <div className="relative flex flex-row gap-4">
+      <SettingsPanel />
+      <div>
+        <Chat />
+      </div>
+    </div>
+  );
+};
+
+export default ChatUI;
