+++ src/pages/gptchat.astro
@@ -1,5 +1,5 @@
 ---
-import ChatUI from '@components/ChatUI.tsx';
+import ChatUI from '@components/ChatUI';
 import BaseLayout from '@layouts/BaseLayout.astro';
 ---
